requirements.txt
# Core dependencies
python-dotenv>=1.0.0
selenium>=4.15.2
undetected-chromedriver>=3.5.4
webdriver-manager>=4.0.1
gradio>=4.13.0
google-generativeai>=0.3.2
crawl4ai>=0.1.0
requests>=2.31.0

# Utility packages
asyncio>=3.4.3
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
pandas>=2.1.4
numpy>=1.26.2

# Logging and monitoring
logging>=0.4.9.6

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.2

# Type checking
mypy>=1.7.1
types-requests>=2.31.0.10
types-beautifulsoup4>=4.12.0.7 